requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
scikit-learn
evidently
dill
-e .
